requirements.txt
asgiref==3.5.1
autopep8==1.6.0
Django==4.0.4
django-ckeditor==6.4.0
django-js-asset==2.0.0
gunicorn==20.1.0
Pillow==9.1.0
psycopg2==2.9.3
pycodestyle==2.8.0
sqlparse==0.4.2
toml==0.10.2
tzdata==2022.1
whitenoise==6.1.0
